[CP-3435] Optimized MenuItem re-renders by using memoization and prop destructuring
